+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\PC_User\.android\avd\Pixel_4_API_30.avd" />
+            <value value="C:\Users\PC_User\.android\avd\Pixel_4_API_28.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-10-14T05:50:38.671899200Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-10-24T12:54:31.589972900Z" />
   </component>
 </project>
