--- app/src/main/java/jp/example/tanmen/Model/Shop.kt
@@ -1,10 +1,0 @@
-package jp.example.tanmen.Model
-
-import java.io.Serializable
-
-data class Shop(
-    var image: String,
-    var name: String,
-    var address: String?,
-    var hours: String?
-): Serializable
