+++ app/src/main/java/jp/example/tanmen/View/Adapter/BottomNavigationPagerAdapter.kt
@@ -1,0 +1,27 @@
+package jp.example.tanmen.View.Adapter
+
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import jp.example.tanmen.View.Fragment.HomeFragment
+import jp.example.tanmen.View.Fragment.MainFragment
+import jp.example.tanmen.View.Fragment.ShuffleFragment
+
+class BottomNavigationPagerAdapter(fm: MainFragment) : FragmentStateAdapter(fm) {
+    override fun getItemCount(): Int {
+        return 2
+    }
+
+    override fun createFragment(position: Int): Fragment {
+        return when(position) {
+            0 -> {
+                HomeFragment()
+            }
+            1 -> {
+                ShuffleFragment()
+            }
+            else -> {
+                HomeFragment()
+            }
+        }
+    }
+}
