--- app/src/main/java/jp/example/tanmen/View/BottomNavigationPagerAdapter.kt
@@ -1,27 +1,0 @@
-package jp.example.tanmen.View
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import jp.example.tanmen.Controller.Fragment.HomeFragment
-import jp.example.tanmen.Controller.Fragment.MainFragment
-import jp.example.tanmen.Controller.Fragment.ShuffleFragment
-
-class BottomNavigationPagerAdapter(fm: MainFragment) : FragmentStateAdapter(fm) {
-    override fun getItemCount(): Int {
-        return 2
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return when(position) {
-            0 -> {
-                HomeFragment()
-            }
-            1 -> {
-                ShuffleFragment()
-            }
-            else -> {
-                HomeFragment()
-            }
-        }
-    }
-}
