+++ app/src/main/res/layout/activity_main.xml
@@ -4,11 +4,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Controller.Activity.MainActivity">
+    tools:context=".View.Activity.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment"
-        android:name="jp.example.tanmen.Controller.Fragment.MainFragment"
+        android:name="jp.example.tanmen.View.Fragment.MainFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintEnd_toEndOf="parent"
