+++ app/src/main/res/layout/fragment_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".Controller.Fragment.DetailFragment">
+    tools:context=".View.Fragment.DetailFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
