+++ app/src/main/res/layout/fragment_shuffle.xml
@@ -1,73 +1,82 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Controller.Fragment.ShuffleFragment">
+<layout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
 
-    <LinearLayout
+    <data>
+        <variable
+            name="shuffleViewModel"
+            type="jp.example.tanmen.ViewModel.ShuffleViewModel" />
+    </data>
+
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:paddingBottom="24dp">
+        android:layout_height="match_parent"
+        tools:context=".View.Fragment.ShuffleFragment">
 
-        <ImageView
-            android:id="@+id/shop_photo"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
-            android:scaleType="fitCenter"
-            android:layout_marginTop="20dp"
-            android:src="@drawable/ramen_dining_24px"
-            android:layout_gravity="center" />
-
-        <TextView
-            android:id="@+id/name_label"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:paddingBottom="8dp"
-            android:textColor="@color/black"
-            android:textSize="20sp"
-            android:textStyle="bold"
-            android:text="店名"
-            android:visibility="gone"
-            android:layout_marginStart="50dp"/>
+            android:orientation="vertical"
+            android:paddingBottom="24dp">
 
-        <TextView
-            android:id="@+id/shop_name"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:text=""
-            android:textColor="@android:color/black"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            android:layout_gravity="center"/>
+            <ImageView
+                android:id="@+id/shop_photo"
+                android:layout_width="300dp"
+                android:layout_height="300dp"
+                android:scaleType="fitCenter"
+                android:layout_marginTop="20dp"
+                android:src="@drawable/ramen_dining_24px"
+                android:layout_gravity="center" />
 
-        <TextView
-            android:id="@+id/address_label"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:paddingBottom="8dp"
-            android:text="住所"
-            android:visibility="gone"
-            android:textColor="@color/black"
-            android:textSize="20sp"
-            android:textStyle="bold"
-            android:layout_marginStart="50dp"/>
+            <TextView
+                android:id="@+id/name_label"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:paddingBottom="8dp"
+                android:textColor="@color/black"
+                android:textSize="20sp"
+                android:textStyle="bold"
+                android:text="店名"
+                android:visibility="gone"
+                android:layout_marginStart="50dp"/>
 
-        <TextView
-            android:id="@+id/shop_address"
-            android:layout_width="300dp"
-            android:layout_height="wrap_content"
-            android:text=""
-            android:textColor="@android:color/black"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            android:layout_gravity="center"/>
+            <TextView
+                android:id="@+id/shop_name"
+                android:layout_width="300dp"
+                android:layout_height="wrap_content"
+                android:text="@={shuffleViewModel.shopName}"
+                android:textColor="@android:color/black"
+                android:textSize="18sp"
+                android:textStyle="bold"
+                android:layout_gravity="center"/>
+
+            <TextView
+                android:id="@+id/address_label"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:paddingBottom="8dp"
+                android:text="住所"
+                android:visibility="gone"
+                android:textColor="@color/black"
+                android:textSize="20sp"
+                android:textStyle="bold"
+                android:layout_marginStart="50dp"/>
+
+            <TextView
+                android:id="@+id/shop_address"
+                android:layout_width="300dp"
+                android:layout_height="wrap_content"
+                android:text="@={shuffleViewModel.shopAddress}"
+                android:textColor="@android:color/black"
+                android:textSize="18sp"
+                android:textStyle="bold"
+                android:layout_gravity="center"/>
 
 
 
-    </LinearLayout>
-</ScrollView>
+        </LinearLayout>
+    </ScrollView>
+</layout>
